backend/requirements.txt
fastapi
uvicorn
motor
pydantic
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[bcrypt]
emails
python-dotenv
aiofiles
opencv-python-headless
tensorflow
numpy
Pillow
requests
